Now will correctly apply added delay
Changed project structure to work from source folder
