--- src/HighlightedItems.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HighlightedItems", "HighlightedItems\HighlightedItems.csproj", "{6D997C8E-5C70-4068-844B-E8BD951C9F42}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6D997C8E-5C70-4068-844B-E8BD951C9F42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6D997C8E-5C70-4068-844B-E8BD951C9F42}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6D997C8E-5C70-4068-844B-E8BD951C9F42}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6D997C8E-5C70-4068-844B-E8BD951C9F42}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7A23B2E7-15CD-4AD9-8B2E-D0826F0CD9D3}
-	EndGlobalSection
-EndGlobal
